Data is stored as json file
The dictionary used to be made every time the page was loaded. The data is stored as json now when the save_details(id) funstion is called. This function is to be later used to revise the data once functionality is added for users to add and remove books. The json files are also stored locally. This can later be changed by integrating with present database